requirements.txt
# Web framework
flask==3.0.0
flask-cors==4.0.0

# Data APIs
requests==2.31.0
feedparser==6.0.11
finnhub-python==2.4.20

# Data processing
numpy==2.1.0
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
hmmlearn==0.3.2

# NLP
vaderSentiment==3.3.2

# UI
streamlit==1.38.0
plotly==5.22.0

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
diskcache==5.6.3

# Testing
pytest==8.3.2
pytest-mock==3.14.0
